+++ .eslintrc.json
@@ -7,8 +7,7 @@
     "next",
     "next/core-web-vitals",
     "plugin:@typescript-eslint/recommended",
-    "prettier",
-    "plugin:storybook/recommended"
+    "prettier"
   ],
   "plugins": ["@typescript-eslint"],
   "root": true,
