--- .storybook/preview.js
@@ -1,25 +1,0 @@
-import 'utils/prototype';
-import 'assets/scss/foundations/base.scss';
-import 'assets/scss/foundations/reset.scss';
-import 'assets/scss/foundations/global.scss';
-import { RouterContext } from "next/dist/shared/lib/router-context";
-
-export const parameters = {
-  actions: { argTypesRegex: "^on[A-Z].*" },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-  nextRouter: {
-    Provider: RouterContext.Provider,
-    nextRouter: {
-      pathname: '/',
-      asPath: '/',
-      query: {},
-      push(url, as, option = undefined) {
-      }
-    }
-  },
-}
