+++ next.config.js
@@ -3,6 +3,9 @@
 // pwa
 const withPWA = require('next-pwa');
 const runtimeCaching = require('next-pwa/cache');
+const withBundleAnalyzer = require('@next/bundle-analyzer')({
+  enabled: process.env.ANALYZE === 'true',
+});
 
 const cacheHeaders = [
   {
@@ -101,4 +104,4 @@ const nextConfig = {
   },
 };
 
-module.exports = withPWA(nextConfig);
+module.exports = withBundleAnalyzer(withPWA(nextConfig));
