+++ package.json
@@ -8,9 +8,11 @@
   },
   "scripts": {
     "dev": "run-p dev:*",
-    "dev:front": "npx rimraf .next && vercel dev -l 1111",
+    "dev:front": "run-s prebuild && vercel dev -l 1111",
     "dev:back": "cd ../MeLoN && npm run start:dev",
-    "build": "npx rimraf .next && next build && build-storybook -o public/_storybook",
+    "build": "run-s prebuild && next build",
+    "analyze": "run-s prebuild && ANALYZE=true next build",
+    "prebuild": "npx rimraf .next",
     "start": "next start",
     "lint": "next lint",
     "format": "npx prettier --write \"src/**/*.ts\" \"src/**/*.tsx\" \"src/**/*.scss\"",
@@ -18,14 +20,13 @@
     "vercel": "vercel dev -l 1111",
     "generate": "graphql-codegen --config codegen.yml",
     "cy:open": "cypress open",
-    "cy:run": "cypress run",
-    "sb:dev": "start-storybook -p 6006",
-    "sb:build": "build-storybook -o public/_storybook"
+    "cy:run": "cypress run"
   },
   "dependencies": {
     "@apollo/client": "^3.5.6",
     "@iconify/react": "^3.1.0",
     "@monaco-editor/react": "^4.3.1",
+    "@next/bundle-analyzer": "^12.0.10",
     "@octokit/rest": "^18.12.0",
     "@stackblitz/sdk": "^1.5.3",
     "axios": "^0.25.0",
@@ -45,6 +46,7 @@
     "react-slick": "^0.28.1",
     "react-split": "^2.0.13",
     "react-swipeable": "^6.2.0",
+    "react-zlib-js": "^1.0.5",
     "swr": "^1.1.0"
   },
   "devDependencies": {
@@ -54,12 +56,6 @@
     "@graphql-codegen/typescript": "2.4.2",
     "@graphql-codegen/typescript-operations": "2.2.2",
     "@graphql-codegen/typescript-react-apollo": "3.2.4",
-    "@storybook/addon-actions": "^6.4.15",
-    "@storybook/addon-essentials": "^6.4.15",
-    "@storybook/addon-links": "^6.4.15",
-    "@storybook/builder-webpack5": "^6.4.15",
-    "@storybook/manager-webpack5": "^6.4.15",
-    "@storybook/react": "^6.4.15",
     "@testing-library/cypress": "^8.0.2",
     "@types/diff": "^5.0.2",
     "@types/gtag.js": "^0.0.8",
@@ -78,14 +74,12 @@
     "eslint": "^7.23.0",
     "eslint-config-next": "^12.0.4",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-storybook": "^0.5.6",
     "monaco-editor": "^0.31.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.5.0",
     "rimraf": "^3.0.2",
     "sass": "^1.49.0",
     "sass-loader": "^12.4.0",
-    "storybook-addon-next-router": "^3.1.1",
     "style-loader": "^3.3.1",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
     "typescript": "^4.5.5"
