+++ src/types/modules.d.ts
@@ -1,0 +1,4 @@
+declare module 'react-zlib-js' {
+  function gzipSync(buf: InputType, options?: ZlibOptions): Buffer;
+  function unzipSync(buf: InputType, options?: ZlibOptions): Buffer;
+}
